chore: upgrade to Spring Boot 3
Spring Boot 3 made some changes to the way a library, such as this one,
notifies Spring Boot that it wants to be autoconfigured.

[Details of this change are in the release notes of version 2.7][1].

With this commit, a new versioning system for this package is proposed:
`<Spring Boot version>-<tradition semantic versioning>`. For example,
`3-1.0.2`. Why? The problem with only semantic versioning _and_
following Spring releases (especially those with breaking changes) is,
that it becomes unclear when to update to a new major version. Should we
do it when this library introduces a breaking change? Or when Spring
Boot does? To solve this, the version now consists of a leading digit,
indicating the Spring Boot version that this version works with,
followed by a semantic version of the library itself.

So, we now have `2-1.0.2`, working with Spring Boot 2, alongside
`3-1.0.2`, which works with Spring Boot 3. When a breaking change is
made to this library, we would signal that by bumping the versions to
`2-2.0.0` and `3-2.0.0`, respectively.

[1]: <https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.7-Release-Notes#changes-to-auto-configuration>